requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
opencv-python>=4.9.0
numpy>=1.24.0
scikit-image>=0.21.0
pillow>=10.0.0
tqdm>=4.66.0
albumentations>=1.3.0
matplotlib>=3.7.0
pyyaml>=6.0
einops>=0.7.0
timm>=0.9.0
moviepy>=1.0.3
ffmpeg-python>=0.2.0

# Optional: For MPRNet (install if you have MPRNet weights)
# basicsr>=1.4.2
# facexlib>=0.2.5

# Optional: For Real-ESRGAN (install if you have Real-ESRGAN weights)
# realesrgan>=0.2.5.0

